refactor(app): modularize MiddlePaneList and optimize rendering performance
- Split giant MiddlePaneList component into subcomponents to isolate tab state and reduce re-renders.
- Provide instantaneous 0ms feedback for tab switching independently from URL query parameter updates.
- Prevent unintended search filter resets during note detail navigation and pin state toggles.
- Apply React.memo to NoteItem to minimize re-rendering overhead during text input and filtering.
- Update bulk copy logic to target currently displayed items instead of auto-excluding resolved notes.
